requirements.txt
alembic
crypto 
cryptography
cybox
decorator
flask
flask-migrate
flask-sqlalchemy
gitdb
gitpython
httplib2
jbxapi 
lief
maec
mako
marshmallow
marshmallow-sqlalchemy
mixbox
neo
oauth
olefile
otx-misp
pdftools 
pefile 
py2neo
plyara
psycopg2-binary
pycrypto
pydeep
pylint
pymisp
pymisp[fileobjects,openioc,virustotal,pdfexport]
python-dateutil
python-dotenv
python-magic
pytz
pyyaml
pyzmq
redis
requests
six>=1.15.0
secretstorage
setuptools
sqlalchemy 
sqlalchemy-utils
ssdeep
stix
stix2
stix2-patterns
toml
urllib3
validators
virustotal-api 
wheel
wrapt
zmq
